+++ test/include/test_dlai_main.h
@@ -1,0 +1,11 @@
+#ifndef DLA_INTERFACE_TEST_INCLUDE_TEST_DLAI_MAIN_H
+#define DLA_INTERFACE_TEST_INCLUDE_TEST_DLAI_MAIN_H
+
+#include <fstream>
+#include <vector>
+#include "communicator_grid.h"
+
+extern std::vector<dla_interface::comm::Communicator2DGrid*> comms;
+extern std::ofstream* outstream;
+
+#endif  // DLA_INTERFACE_TEST_INCLUDE_TEST_DLAI_MAIN_H
